Split the list
In merge-sort, the list will be splited for sorting.In first iteration,
take two elements from the list to sort and the two sorted elements is
added into tmp_list.The above step is implemented until the original
list is empty.sorted_1_list and sorted_2_list are two temporary lists
to link the taken-away nodes.When the original list is empty,if some
elements are not compared and merged, they should be added into the
last merged list and the new list should be sorted,which is implemented
by merge_tail() but merge_tail() isn't written.After adding all elements
into tmp_list, splice tmp_list into unsorted_list and initialize
tmp_list.In second iteration,the program will take four element from
unsorted_list and do the same as above.In the nth iteration, the
program will take n square of two's elements.

n is used to record the times to take elements into sorted_1_list,
twice_n is used to record the times to take elements into
sorted_2_list.In the third condition of the for loop,when compare and
merge, i is still added one, but there is not an element to be taken,
so i should be substracted with one.