Merge list form tail
In merge_tail(),no_merge_list is the remained list that has no parner
to compare and merge.The remained list should be added into the last
merged list.Since the last merged list is in the end of tmp_list,so
inserted the elements of the remained list from the tail of tmp_list.
The new merged list should be sorted,so use the method of insertion
sort to insert.